move from migrate to install the import of data
